update drawer list colors
