+++ lib/components/app_drawer/drive_list_tile.dart
@@ -1,0 +1,18 @@
+import 'package:ardrive/models/models.dart';
+import 'package:flutter/material.dart';
+
+class DriveListTile extends StatelessWidget {
+  final Drive drive;
+  final bool selected;
+  final VoidCallback onPressed;
+
+  const DriveListTile({this.drive, this.selected = false, this.onPressed});
+
+  @override
+  Widget build(BuildContext context) => ListTile(
+        leading: Icon(drive.isPrivate ? Icons.folder : Icons.folder_shared),
+        title: Text(drive.name),
+        selected: selected,
+        onTap: onPressed,
+      );
+}
