+++ lib/components/components.dart
@@ -1,5 +1,5 @@
 export 'app_dialog.dart';
-export 'app_drawer.dart';
+export 'app_drawer/app_drawer.dart';
 export 'confirmation_dialog.dart';
 export 'drive_attach_form.dart';
 export 'drive_create_form.dart';
